requirements.txt
dnspython==2.7.0
kafka-python==2.2.15
pymongo==4.13.2
PySide6
PySide6-Addons
PySide6-Essentials
pyspark==3.5.1
shiboken6
bcrypt==4.3.0
pandas
python-dotenv
PyArrow